meeting-intelligence-agent/backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Task Queue
celery==5.3.6
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.1

# AI & ML
openai==1.10.0
anthropic==0.8.1
tiktoken==0.5.2
langchain==0.1.4
langchain-openai==0.0.2

# Transcription
openai-whisper==20231117
faster-whisper==0.10.0
torch==2.1.2
torchaudio==2.1.2

# Speaker Diarization
pyannote.audio==3.1.1
pyannote.core==5.0.0

# NLP & Text Processing
spacy==3.7.2
transformers==4.37.0
sentence-transformers==2.3.1
nltk==3.8.1

# Vector Store
pinecone-client==3.0.2
chromadb==0.4.22

# Integrations - Calendar
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0
O365==2.0.27

# Integrations - Communication
slack-sdk==3.26.2
python-telegram-bot==20.7
resend==0.7.0

# Integrations - Video Platforms
zoom-python==0.0.2
microsoft-teams-python==0.1.1

# Integrations - Project Management
jira==3.6.0
asana==4.0.7
notion-client==2.2.1

# WebSocket
python-socketio==5.11.0
websockets==12.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Data Validation & Serialization
marshmallow==3.20.2
orjson==3.9.12

# File Processing
python-magic==0.4.27
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0

# Monitoring & Logging
prometheus-client==0.19.0
sentry-sdk==1.40.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.2.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2024.1
tenacity==8.2.3
